+++ src/http/api/routes/resolve.js
@@ -1,0 +1,16 @@
+'use strict'
+
+const resources = require('./../resources')
+
+module.exports = (server) => {
+  const api = server.select('API')
+
+  api.route({
+    method: '*',
+    path: '/api/v0/resolve',
+    config: {
+      handler: resources.resolve.handler,
+      validate: resources.resolve.validate
+    }
+  })
+}