+++ test/http-api/interface.js
@@ -25,28 +25,7 @@ describe('interface-ipfs-core over ipfs-api tests', () => {
     skip: { reason: 'TODO: DHT is not implemented in js-ipfs yet!' }
   })
 
-  tests.files(defaultCommonFactory, {
-    skip: [
-      // files.ls
-      {
-        name: 'should ls directory',
-        reason: 'FIXME: https://github.com/ipfs/js-ipfs-mfs/issues/7'
-      },
-      {
-        name: 'should ls -l directory',
-        reason: 'FIXME: https://github.com/ipfs/js-ipfs-mfs/issues/7'
-      },
-      // files.read*Stream
-      {
-        name: 'readPullStream',
-        reason: 'FIXME: https://github.com/ipfs/js-ipfs-mfs/issues/8'
-      },
-      {
-        name: 'readReadableStream',
-        reason: 'FIXME: https://github.com/ipfs/js-ipfs-mfs/issues/8'
-      }
-    ]
-  })
+  tests.files(defaultCommonFactory)
 
   tests.key(CommonFactory.create({
     spawnOptions: {
@@ -61,8 +40,12 @@ describe('interface-ipfs-core over ipfs-api tests', () => {
   }), {
     skip: [
       {
-        name: 'resolve',
-        reason: 'TODO: not implemented'
+        name: 'should resolve an IPNS DNS link',
+        reason: 'TODO IPNS not implemented yet'
+      },
+      {
+        name: 'should resolve IPNS link recursively',
+        reason: 'TODO IPNS not implemented yet'
       }
     ]
   })